Level API
API for level

+++ backend/src/main/java/org/siges/backend/dao/SectionRepository.java
@@ -3,6 +3,6 @@ package org.siges.backend.dao;
 import org.siges.backend.entity.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectionRepository extends JpaRepository<Section, Long> {
+public interface SectionRepository extends JpaRepository<Section, Integer> {
 
 }
